chore(test-project): Reuse api tasks between test-project and rebuild-test-project-fixture
