+++ tasks/test-project/tui-tasks.mts
@@ -1,25 +1,8 @@
-import fs from 'node:fs'
-import path from 'node:path'
-
 import type { Options as ExecaOptions } from 'execa'
 
-import {
-  createBuilder,
-  fullPath,
-  getOutputPath,
-  setOutputPath,
-  addModel,
-  addDbAuth,
-  webTasksList,
-} from './base-tasks.mts'
-import { getPrerenderTasks } from './prerender-tasks.mts'
+import { setOutputPath, webTasksList } from './base-tasks.mts'
 import type { TuiTaskList } from './typing.mts'
-import {
-  getExecaOptions as utilGetExecaOptions,
-  applyCodemod,
-  exec,
-  getCfwBin,
-} from './util.mts'
+import { getExecaOptions as utilGetExecaOptions, exec } from './util.mts'
 
 function getExecaOptions(cwd: string): ExecaOptions {
   return { ...utilGetExecaOptions(cwd), stdio: 'pipe' as const }
@@ -38,276 +21,7 @@ export async function webTasks(outputPath: string) {
         await exec('yarn cedar setup ui tailwindcss', ['--force'], execaOptions)
       },
     },
-  ] //,
-  // TODO: Figure out what to do with this. It's from Listr, but TUI doesn't
-  //       have anything like it (yet?)
-  // {
-  //   exitOnError: true,
-  //   renderer: verbose && 'verbose',
-  // }
-
-  return tuiTaskList
-}
-
-interface ApiTasksOptions {
-  linkWithLatestFwBuild: boolean
-  esmProject: boolean
-}
-
-export async function apiTasks(
-  outputPath: string,
-  { linkWithLatestFwBuild, esmProject }: ApiTasksOptions,
-) {
-  setOutputPath(outputPath)
-
-  const execaOptions = getExecaOptions(outputPath)
-  const generateScaffold = createBuilder('yarn cedar g scaffold')
-
-  const tuiTaskList: TuiTaskList = [
-    {
-      title: 'Adding post and user model to prisma',
-      task: async () => {
-        // Need both here since they have a relation
-        const { post, user } = await import('./codemods/models.mts')
-
-        addModel(post)
-        addModel(user)
-
-        return exec(
-          `yarn cedar prisma migrate dev --name create_post_user`,
-          [],
-          execaOptions,
-        )
-      },
-    },
-    {
-      title: 'Scaffolding post',
-      task: async () => {
-        await generateScaffold('post')
-
-        // Replace the random numbers in the scenario with consistent values
-        await applyCodemod(
-          'scenarioValueSuffix.js',
-          fullPath('api/src/services/posts/posts.scenarios'),
-        )
-
-        await exec(
-          `yarn ${getCfwBin(getOutputPath())} project:copy`,
-          [],
-          execaOptions,
-        )
-      },
-    },
-    {
-      title: 'Adding seed script',
-      task: async () => {
-        await applyCodemod(
-          'seed.js',
-          fullPath('scripts/seed.ts', { addExtension: false }),
-        )
-      },
-    },
-    {
-      title: 'Adding contact model to prisma',
-      task: async () => {
-        const { contact } = await import('./codemods/models.mts')
-
-        addModel(contact)
-
-        await exec(
-          `yarn cedar prisma migrate dev --name create_contact`,
-          [],
-          execaOptions,
-        )
-
-        await generateScaffold('contacts')
-
-        const contactsServicePath = fullPath(
-          'api/src/services/contacts/contacts',
-        )
-        fs.writeFileSync(
-          contactsServicePath,
-          fs
-            .readFileSync(contactsServicePath, 'utf-8')
-            .replace(
-              "import { db } from 'src/lib/db'",
-              '// Testing aliased imports with extensions\n' +
-                "import { db } from 'src/lib/db.js'",
-            ),
-        )
-      },
-    },
-    {
-      // This task renames the migration folders so that we don't have to deal
-      // with duplicates/conflicts when committing to the repo
-      title: 'Adjust dates within migration folder names',
-      task: () => {
-        const migrationsFolderPath = path.join(
-          getOutputPath(),
-          'api',
-          'db',
-          'migrations',
-        )
-        // Migration folders are folders which start with 14 digits because they
-        // have a yyyymmddhhmmss
-        const migrationFolders = fs
-          .readdirSync(migrationsFolderPath)
-          .filter((name) => {
-            return (
-              name.match(/\d{14}.+/) &&
-              fs.lstatSync(path.join(migrationsFolderPath, name)).isDirectory()
-            )
-          })
-          .sort()
-        const datetime = new Date('2022-01-01T12:00:00.000Z')
-        migrationFolders.forEach((name) => {
-          const datetimeInCorrectFormat =
-            datetime.getFullYear() +
-            ('0' + (datetime.getMonth() + 1)).slice(-2) +
-            ('0' + datetime.getDate()).slice(-2) +
-            '120000' // Time hardcoded to 12:00:00 to limit TZ issues
-          fs.renameSync(
-            path.join(migrationsFolderPath, name),
-            path.join(
-              migrationsFolderPath,
-              `${datetimeInCorrectFormat}${name.substring(14)}`,
-            ),
-          )
-          datetime.setDate(datetime.getDate() + 1)
-        })
-      },
-    },
-    {
-      title: 'Add dbAuth',
-      task: async () => addDbAuth(outputPath, linkWithLatestFwBuild),
-    },
-    {
-      title: 'Add users service',
-      task: async () => {
-        const generateSdl = createBuilder('yarn cedar g sdl --no-crud', 'api')
-
-        await generateSdl('user')
-
-        await applyCodemod('usersSdl.js', fullPath('api/src/graphql/users.sdl'))
-
-        await applyCodemod(
-          'usersService.js',
-          fullPath('api/src/services/users/users'),
-        )
-
-        // Replace the random numbers in the scenario with consistent values
-        await applyCodemod(
-          'scenarioValueSuffix.js',
-          fullPath('api/src/services/users/users.scenarios'),
-        )
-
-        const test = `import { user } from './users.js'
-            import type { StandardScenario } from './users.scenarios.js'
-
-            describe('users', () => {
-              scenario('returns a single user', async (scenario: StandardScenario) => {
-                const result = await user({ id: scenario.user.one.id })
-
-                expect(result).toEqual(scenario.user.one)
-              })
-            })`.replaceAll(/ {12}/g, '')
-
-        fs.writeFileSync(fullPath('api/src/services/users/users.test'), test)
-
-        return createBuilder('yarn cedar g types')()
-      },
-    },
-    {
-      title: 'Add describeScenario tests',
-      task: () => {
-        // Copy contact.scenarios.ts, because scenario tests look for the same filename
-        fs.copyFileSync(
-          fullPath('api/src/services/contacts/contacts.scenarios'),
-          fullPath('api/src/services/contacts/describeContacts.scenarios'),
-        )
-
-        // Create describeContacts.test.ts
-        const describeScenarioFixture = path.join(
-          import.meta.dirname,
-          'templates',
-          'api',
-          'contacts.describeScenario.test.ts.template',
-        )
-
-        fs.copyFileSync(
-          describeScenarioFixture,
-          fullPath('api/src/services/contacts/describeContacts.test'),
-        )
-      },
-    },
-    {
-      // This is probably more of a web side task really, but the scaffolded
-      // pages aren't generated until we get here to the api side tasks. So
-      // instead of doing some up in the web side tasks, and then the rest
-      // here I decided to move all of them here
-      title: 'Add Prerender to Routes',
-      task: async () => getPrerenderTasks(),
-    },
-    {
-      title: 'Add context tests',
-      task: () => {
-        const templatePath = path.join(
-          import.meta.dirname,
-          'templates',
-          'api',
-          'context.test.ts.template',
-        )
-        const projectPath = path.join(
-          getOutputPath(),
-          'api',
-          'src',
-          '__tests__',
-          'context.test.ts',
-        )
-
-        fs.mkdirSync(path.dirname(projectPath), { recursive: true })
-        fs.writeFileSync(projectPath, fs.readFileSync(templatePath))
-      },
-    },
-    {
-      title: 'Add vitest db import tracking tests for ESM test project',
-      task: () => {
-        if (!esmProject) {
-          return
-        }
-
-        const templatesDir = path.join(import.meta.dirname, 'templates', 'api')
-        const templatePath1 = path.join(templatesDir, '1-db-import.test.ts')
-        const templatePath2 = path.join(templatesDir, '2-db-import.test.ts')
-        const templatePath3 = path.join(templatesDir, '3-db-import.test.ts')
-
-        const testsDir = path.join(getOutputPath(), 'api', 'src', '__tests__')
-        const testFilePath1 = path.join(testsDir, '1-db-import.test.ts')
-        const testFilePath2 = path.join(testsDir, '2-db-import.test.ts')
-        const testFilePath3 = path.join(testsDir, '3-db-import.test.ts')
-
-        fs.mkdirSync(testsDir, { recursive: true })
-        fs.copyFileSync(templatePath1, testFilePath1)
-        fs.copyFileSync(templatePath2, testFilePath2)
-        fs.copyFileSync(templatePath3, testFilePath3)
-
-        // I opted to add an additional vitest config file rather than modifying
-        // the existing one because I wanted to keep one looking exactly the
-        // same as it'll look in user's projects.
-        fs.copyFileSync(
-          path.join(templatesDir, 'vitest-sort.config.ts'),
-          path.join(getOutputPath(), 'api', 'vitest-sort.config.ts'),
-        )
-      },
-    },
   ]
-  // ],
-  // TODO: Figure out what to do with this. It's from Listr, but TUI doesn't
-  //       have anything like it (yet?)
-  // {
-  //   exitOnError: true,
-  //   renderer: verbose && 'verbose',
-  //   renderOptions: { collapseSubtasks: false },
-  // }
+
   return tuiTaskList
 }
